+++ packages/MagicImpute/src/founderimpute.jl
@@ -9,53 +9,64 @@ function founderimpute_chr!(chrfhaplo::AbstractMatrix, chroffgeno::AbstractMatri
     allelebalancemean::Union{Real,AbstractVector},
     allelebalancedisperse::Union{Real,AbstractVector},
     alleledropout::Union{Real,AbstractVector},
-    offspringexcl::AbstractVector, 
-    inputloglike::Real,
+    offspringexcl::AbstractVector,
     snporder::AbstractVector,
     israndallele::Bool,     
     issnpGT::AbstractVector,
     upbyhalf::Bool,
-    isallowmissing::Bool=false,
+    alwaysaccept::Bool, 
+    isallowmissing::Bool=true,
     imputetempfile::AbstractString)    
     nsnp = size(chrfhaplo,1)
     if nsnp == 1
         @warn string("TODO: nsnp=",nsnp)
     end
-    nsnp <=1 && return chrfhaplo
+    nsnp <=1 && return (chrfhaplo, 0, 0)
     nsnp == length(snporder) || @error "dimension mismatch!"
+    loglikels = MagicReconstruct.hmm_loglikels(chrfhaplo,chroffgeno,
+        popmakeup,priorprocess;
+        epsf,epso,epso_perind, seqerror,
+        allelebalancemean,allelebalancedisperse,alleledropout, 
+        decodetempfile=imputetempfile, israndallele,issnpGT,snporder
+    )	
+    loglikels[offspringexcl] .= 0.0 
+    inputloglike = sum(loglikels)
+    
     isfounderinbred = length(fhaplosetpp) == size(chrfhaplo,2)    
     newchrfhaplo = deepcopy(chrfhaplo)
     snpincl = snporder[first(values(priorprocess)).markerincl]				    
-    if upbyhalf            
-        for reversechr in [false,true]                
-            # fixing phasing on one half of the chromosome                      
-            nhalf = nsnp ÷ 2                
-            fhaplosetpp2 = deepcopy(fhaplosetpp)                
-            snpls = reversechr ? snporder[1:nhalf] : snporder[nhalf+1:nsnp]
-            for p in eachindex(fhaplosetpp2)                    
-                if isfounderinbred
-                    for j in snpls
-                        fhaplosetpp2[p][j] = [newchrfhaplo[j,p]]                        
+    if upbyhalf                    
+        for findex in findexlist        
+            for reversechr in [false,true]                      
+                # fixing phasing on one half of the chromosome                      
+                nhalf = nsnp ÷ 2                
+                fhaplosetpp2 = deepcopy(fhaplosetpp)                
+                snpls = reversechr ? snporder[1:nhalf] : snporder[nhalf+1:nsnp]
+                for p in eachindex(fhaplosetpp2)                    
+                    if isfounderinbred
+                        for j in snpls
+                            fhaplosetpp2[p][j] = [newchrfhaplo[j,p]]                        
+                        end
+                    else
+                        for j in snpls
+                            fhaplosetpp2[p][j] = [newchrfhaplo[j,[2p-1,2p]]]                        
+                        end
                     end
-                else
-                    for j in snpls
-                        fhaplosetpp2[p][j] = [newchrfhaplo[j,[2p-1,2p]]]                        
-                    end
-                end
-            end
-            for findex in findexlist  
+                end                
                 founderforwardbackward!(findex,
                     newchrfhaplo,chroffgeno,
                     popmakeup,priorprocess,fhaplosetpp2; 
                     epsf,epso,epso_perind,seqerror,
                     allelebalancemean,allelebalancedisperse,alleledropout,
                     offspringexcl, snporder,israndallele, issnpGT,
-                    reversechr = !reversechr, forwardtempfile=imputetempfile, isallowmissing)            
+                    reversechr = !reversechr, forwardtempfile=imputetempfile, 
+                    isallowmissing
+                )            
             end 
         end
-    else
-        reversechr =  rand([false,true])
-        for findex in findexlist                    
+    else    
+        for findex in findexlist                 
+            reversechr =  rand([false,true])   
             founderforwardbackward!(findex,newchrfhaplo,chroffgeno,
                 popmakeup,priorprocess,fhaplosetpp; 
                 epsf,epso,epso_perind, seqerror,
@@ -72,19 +83,17 @@ function founderimpute_chr!(chrfhaplo::AbstractMatrix, chroffgeno::AbstractMatri
     )	
     loglikels[offspringexcl] .= 0.0 
     newloglike = sum(loglikels)
-    isaccept = newloglike > inputloglike    
+    deltloglike = newloglike - inputloglike    
+    isaccept = alwaysaccept ? true : deltloglike > 0.0
     if isaccept
-        deltloglike = newloglike - inputloglike
         chrfhaplo .= newchrfhaplo
-    else
-        deltloglike = 0.0
     end
     deltloglike, ndiff
 end
 
 function get_ndiff(oldchrfhaplo, newchrfhaplo, snpincl)    
-    oldchrfhaplo2 = deepcopy(oldchrfhaplo[snpincl,:])
-    chrfhaplo2 = view(newchrfhaplo, snpincl,:)
+    oldchrfhaplo2 = oldchrfhaplo[snpincl,:]
+    chrfhaplo2 = newchrfhaplo[snpincl,:]
     perm = MagicBase.permfounder(oldchrfhaplo2, chrfhaplo2)
     if perm != 1:length(perm)
         oldchrfhaplo2 .= oldchrfhaplo2[:,perm]
@@ -287,7 +296,7 @@ function founderforwardbackward!(findex::AbstractVector,
     israndallele::Bool,
     issnpGT::AbstractVector, 
     reversechr::Bool,
-    isallowmissing::Bool=false,    
+    isallowmissing::Bool=true,    
     forwardtempfile::AbstractString)
     if reversechr
         reverse!(snporder)