+++ CHANGELOG.md
@@ -13,6 +13,9 @@
 * MagicBase: update parsebreed for an input breeder's pedfile with some sample indviduals acting also as founders in pedcode
 * MagicReconstruct: new option "thincm"
 
-## RABBIT v1.2.0
-* MagicImpute: additional likelihood-based accept/reject for foundercorrect
+## RABBIT v1.2.1
+
 * MagicFilter and others: rename options by removing "snp_". For example, change snp_minmaf to minmaf
+* MagicImpute: additional likelihood-based accept/reject for foundercorrect
+* MagicImpute: delay accet/reject for founderimpute
+* MagicImpute: reset default isallowmissing = true